backend/servicio_analytics/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy>=1.11.0
django-health-check==3.18.1

# Forecasting & Machine Learning (Prophet)
prophet>=1.1  # Note: fbprophet renamed to prophet in v1.1+
cmdstanpy>=1.0.0

# Monitoring & Observability
prometheus-client==0.19.0
python-json-logger==2.0.7
psutil==5.9.6

# Rate Limiting & Security
django-ratelimit==4.1.0
